pyproject.toml
[project]
name = "ella-fab4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "ipykernel>=7.2.0",
    "lancedb>=0.30.2",
    "mlflow>=3.11.1",
    "pandas>=2.3.3",
    "pydantic>=2.13.3",
    "pydantic-ai>=1.86.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
